pyproject.toml
[tool.poetry]
name = "art2s3"
version = "0.2.0"
description = ""
authors = ["Gerald Manipon <pymonger@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
dohq-artifactory = "^0.7.574"
smart-open = {extras = ["s3"], version = "^5.0.0"}
boto3 = "^1.17.73"
click = "^8.0.0"
pytz = "^2021.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
art2s3 = "art2s3.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
